+++ drift/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.2
+
+- Revert the breaking change around `QueryRow.read` only return non-nullable
+  values now - it was causing issues with type inference in some cases.
+
 ## 2.0.1
 
 - Fix an error when inserting a null value into a nullable column defined with